pywww/requirements.txt
apparmor==3.1.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
asttokens==2.0.8
autocommand==2.2.1
backcall==0.2.0
btrfsutil==5.19.1
CacheControl==0.12.11
catfish==4.16.4
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
colorama==0.4.5
commonmark==0.9.1
cryptography==38.0.1
Cython==0.29.32
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
distro==1.7.0
Django==4.1.1
django-extensions==3.2.1
django-import-export==3.0.1
docopt==0.6.2
docutils==0.19
et-xmlfile==1.1.0
executing==1.1.0
Faker==15.1.1
future==0.18.2
gufw==22.4.0
idna==3.4
inflect==6.0.0
ipython==8.5.0
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
jedi==0.18.1
keyutils==0.6
lensfun==0.3.3
LibAppArmor==3.1.1
lightdm-gtk-greeter-settings==1.2.2
lit==14.0.6.dev0
manjaro-sdk==0.1
MarkupPy==1.14
matplotlib-inline==0.1.6
menulibre==2.2.3
more-itertools==8.13.0
msgpack==1.0.4
mugshot==0.4.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
odfpy==1.4.1
openpyxl==3.0.10
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
parso==0.8.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.21.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyGObject==3.42.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.2
pysmbc==1.0.23
python-dateutil==2.8.2
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.11
requests==2.28.1
resolvelib==0.8.1
rich @ file:///build/python-rich/src/rich-12.6.0/dist/rich-12.6.0-py3-none-any.whl
shiboken6==6.3.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
systemd-python==235
tablib==3.2.1
team==1.0
tenacity==8.1.0
tomli==2.0.1
traitlets==5.4.0
trove-classifiers==2022.9.26
typing_extensions==4.3.0
tzdata==2022.4
udiskie==2.4.2
ufw==0.36.1
urllib3==1.26.12
validate-pyproject==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
